+++ packages/terra-date-input/tests/jest/DateInput.test.jsx
@@ -7,6 +7,8 @@ import DateInput from '../../src/DateInput';
 import DateInputField from '../../src/DateInputField';
 import DateInputUtil from '../../lib/DateInputUtil';
 
+jest.mock('uuid/v4', () => () => '00000000-0000-0000-0000-000000000000');
+
 describe('DateInput', () => {
   const defaultRender = <DateInput name="date-input" />;